node base
