+++ src/aioway/_core/iters.py
@@ -5,26 +5,19 @@
 import abc
 import copy
 import dataclasses as dcls
-import textwrap
 import typing
 from collections import abc as cabc
 
 import tensordict as td
 import torch
-from rich import table
 
 from aioway._utils import (
-    AnyDict,
-    AnySet,
-    dag_node_key,
-    dcls_asdict,
     decomp_dcls_members,
-    decomp_replace,
-    render_fcall_str,
-    topo_sort,
     torch_fake_mode,
 )
 from aioway.spaces import Attr, Shape
+
+from .nodes import GraphNode
 
 if typing.TYPE_CHECKING:
     from .procs import IterProc
@@ -37,7 +30,7 @@ def iter_dcls(cls: type):
     return dcls.dataclass(match_args=False)(cls)
 
 
-class Iter[T](cabc.Iterable[T], abc.ABC):
+class Iter[T](cabc.Iterable[T], GraphNode, abc.ABC):
     """
     The class that defines [h]igh level [op]erations.
     It produces iterators that computes the desired batch, represented by the node.
@@ -74,30 +67,6 @@ class Iter[T](cabc.Iterable[T], abc.ABC):
         with torch_fake_mode():
             return next(iter(self))
 
-    def replace(
-        self,
-        function: cabc.Callable[[Iter], Iter],
-        memo: AnyDict[typing.Any, typing.Any] | None = None,
-    ) -> typing.Self:
-        """
-        Replace the current `node` with `function(node)` if it's not `None`,
-        or else recursively invoke `.replace(function)` on `Iter` in `.inputs()`.
-
-        Args:
-            function: The function to apply to the graph.
-            memo: The memo to use if given.
-        """
-
-        def maybe_replace(hop):
-            if not isinstance(hop, Iter):
-                return NotImplemented
-
-            return function(hop)
-
-        result: typing.Any = decomp_replace(self, maybe_replace, memo=memo)
-        assert type(result) == type(self)
-        return result
-
     def rebuild(self):
         """
         Rebuild the current `Iter`. This is useful when you are switching contexts,
@@ -113,50 +82,12 @@ class Iter[T](cabc.Iterable[T], abc.ABC):
     def _rebuild(self) -> typing.Self:
         return copy.copy(self)
 
-    @property
-    def requires_grad(self) -> bool:
-        """
-        Whether or not the `torch.Tensor`s generated by this `Iter` `requires_grad`.
-        If `False`, much like `torch`'s `requires_grad`,
-        gradient does not need to propagate to this node, and we can do a graph break.
-
-        Defaults to checking if any of the `self.deps()` have any `requires_grad`.
-        """
-
-        return any(hop.requires_grad for hop in self.deps())
-
+    @typing.override
     def deps(self) -> cabc.Iterator[Iter]:
         "Decompose `self`, get the immediate dependencies."
 
         for hop in decomp_dcls_members(self, Iter):
             yield hop
-
-    def nodes(self) -> AnySet[Iter]:
-        "Get all the nodes recursively, including `self`."
-
-        nodes = AnySet[Iter](Iter)
-
-        # Add nodes recursively.
-        def _visit_nodes(hop: Iter):
-            if hop in nodes:
-                return
-
-            nodes.add(hop)
-
-            for dep in hop.deps():
-                _visit_nodes(dep)
-
-        _visit_nodes(self)
-        return nodes
-
-    @property
-    def is_source(self) -> bool:
-        "If the `Iter` is not dependent on other `Iter`s, it's considered a source."
-
-        for _ in decomp_dcls_members(self, Iter):
-            return True
-        else:
-            return False
 
     @property
     def size(self) -> int:
@@ -167,48 +98,6 @@ class Iter[T](cabc.Iterable[T], abc.ABC):
         """
 
         return NotImplemented
-
-    def dag(self) -> list[Iter[T]]:
-        "Order the `Iter`s in their topological order."
-
-        dag_nodes = [dag_node_key(hop) for hop in self.nodes()]
-        return topo_sort(dag_nodes)
-
-    def _repr_abbrev(self) -> str:
-        """
-        Perform `repr`, but with `Iter` members abbreviated.
-        """
-
-        members: dict[str, str] = {}
-
-        for key, val in dcls_asdict(self).items():
-
-            # Only perform abbreviation on `Iter`.
-            if isinstance(val, Iter):
-                string = f"{type(val).__name__}(...)"
-            else:
-                string = textwrap.shorten(repr(val), 20)
-
-            members[key] = string
-
-        return render_fcall_str(type(self).__name__, **members)
-
-    def explain(self) -> table.Table:
-        tab = table.Table(title="Logical plan")
-        tab.add_column("Id")
-        tab.add_column("Parents")
-        tab.add_column("Self")
-
-        dag = self.dag()
-        hop_idx = AnyDict[Iter, int]()
-        for idx, hop in enumerate(dag):
-            hop_idx[hop] = idx
-
-        for idx, hop in enumerate(dag):
-            parents = (f"#{hop_idx[dep]}" for dep in hop.deps())
-            tab.add_row(f"#{idx}", ",".join(parents), hop._repr_abbrev())
-
-        return tab
 
 
 @iter_dcls
