+++ src/aioway/_core/nodes.py
@@ -2,10 +2,25 @@
 
 "Metadata for torch operators / functions."
 
+import abc
+import dataclasses as dcls
+import textwrap
 import typing
 from collections import abc as cabc
 
 import torch
+from rich import table
+
+from aioway._utils import (
+    AnyDict,
+    AnySet,
+    dag_node_key,
+    dcls_asdict,
+    decomp_dcls_members,
+    decomp_replace,
+    render_fcall_str,
+    topo_sort,
+)
 
 __all__ = ["TensorInput"]
 
@@ -20,3 +35,128 @@ class TensorInput(typing.Protocol):
         "The tensor operands (inputs to the function)"
 
         raise NotImplementedError
+
+
+@typing.dataclass_transform()
+def node_dcls(cls: type):
+    return dcls.dataclass(match_args=False)(cls)
+
+
+@node_dcls
+class GraphNode(abc.ABC):
+    """
+    `GraphNode` allows us to reconstruct the computation graph traces `torch` resources.
+    When `.deps()` is not empty, the output depends on the inputs,
+    and `.backward()` would propagate to the `.deps()` nodes (if `.requires_grad`).
+    """
+
+    @property
+    def requires_grad(self) -> bool:
+        """
+        Whether or not the `torch.Tensor`s generated by this `Iter` `requires_grad`.
+        If `False`, much like `torch`'s `requires_grad`,
+        gradient does not need to propagate to this node, and we can do a graph break.
+
+        Defaults to checking if any of the `self.deps()` have any `requires_grad`.
+        """
+
+        return any(hop.requires_grad for hop in self.deps())
+
+    @abc.abstractmethod
+    def deps(self) -> cabc.Iterator[GraphNode]:
+        """
+        This defines the prior nodes that `.backward` would propagate to.
+        """
+
+        for node in decomp_dcls_members(self, GraphNode):
+            yield node
+
+    def nodes(self) -> AnySet[GraphNode]:
+        "Get all the nodes recursively, including `self`."
+
+        nodes = AnySet(GraphNode)
+
+        # Add nodes recursively.
+        def _visit_nodes(hop: GraphNode):
+            if hop in nodes:
+                return
+
+            nodes.add(hop)
+
+            for dep in hop.deps():
+                _visit_nodes(dep)
+
+        _visit_nodes(self)
+        return nodes
+
+    @property
+    def is_source(self) -> bool:
+        "If the `GraphNode` is not dependent on other `Iter`s, it's considered a source."
+
+        return bool(list(self.deps()))
+
+    def dag(self) -> list[GraphNode]:
+        "Order the `Iter`s in their topological order."
+
+        dag_nodes = [dag_node_key(node) for node in self.nodes()]
+        return topo_sort(dag_nodes)
+
+    def replace(
+        self,
+        function: cabc.Callable[[GraphNode], GraphNode],
+        memo: AnyDict[typing.Any, typing.Any] | None = None,
+    ) -> typing.Self:
+        """
+        Replace the current `node` with `function(node)` if it's not `None`,
+        or else recursively invoke `.replace(function)` on `Iter` in `.inputs()`.
+
+        Args:
+            function: The function to apply to the graph.
+            memo: The memo to use if given.
+        """
+
+        def maybe_replace(hop):
+            if not isinstance(hop, GraphNode):
+                return NotImplemented
+
+            return function(hop)
+
+        result: typing.Any = decomp_replace(self, maybe_replace, memo=memo)
+        assert type(result) == type(self)
+        return result
+
+    def _repr_abbrev(self) -> str:
+        """
+        Perform `repr`, but with `Iter` members abbreviated.
+        """
+
+        members: dict[str, str] = {}
+
+        for key, val in dcls_asdict(self).items():
+
+            # Only perform abbreviation on `GraphNode`.
+            if isinstance(val, GraphNode):
+                string = f"{type(val).__name__}(...)"
+            else:
+                string = textwrap.shorten(repr(val), 20)
+
+            members[key] = string
+
+        return render_fcall_str(type(self).__name__, **members)
+
+    def explain(self) -> table.Table:
+        tab = table.Table(title="Logical plan")
+        tab.add_column("Id")
+        tab.add_column("Parents")
+        tab.add_column("Self")
+
+        dag = self.dag()
+        node_idx = AnyDict[GraphNode, int]()
+        for idx, node in enumerate(dag):
+            node_idx[node] = idx
+
+        for idx, node in enumerate(dag):
+            parents = (f"#{node_idx[dep]}" for dep in node.deps())
+            tab.add_row(f"#{idx}", ",".join(parents), node._repr_abbrev())
+
+        return tab
